+++ src/vr/locomotion/VRLocomotion.js
@@ -282,9 +282,7 @@ export class VRLocomotion {
     }
   }
   
-
-
-ormalizeAngle(angle) {
+  normalizeAngle(angle) {
     while (angle > Math.PI) angle -= 2 * Math.PI;
     while (angle < -Math.PI) angle += 2 * Math.PI;
     return angle;